# openapi-spec.yaml
openapi: 3.0.3
info:
  title: Job Agents API
  description: |
    A comprehensive API for managing job applications, automated job research, and application processing.
    This system provides endpoints for job management, application tracking, question handling, and automated agent operations.
  version: 1.0.0
  contact:
    name: Job Agents Development Team
  license:
    name: ISC
servers:
  - url: http://localhost:4000/api
    description: Development server
  - url: https://anvaia-site.onrender.com/api
    description: Production server
  - url: /api
    description: Relative path for proxy setups
paths:
...
  /jobs:
    get:
      summary: Get All Jobs
      description: Retrieve all available jobs ordered by most recently scraped
      tags:
        - Jobs
      responses:
        '200':
          description: List of all jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      summary: Create Job
      description: Create a new single job
      tags:
        - Jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreate'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  job:
                    $ref: '#/components/schemas/Job'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - job already exists with conflicting unique fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /jobs/import:
    post:
      summary: Import Jobs
      description: Import multiple jobs with deduplication based on external_link or internal_id
      tags:
        - Jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/JobImport'
      responses:
        '200':
          description: Jobs imported successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  added:
                    type: integer
                    description: Number of jobs successfully added/updated
                    example: 5
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Job:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the job
          example: "123456789"
        internal_id:
          type: string
          nullable: true
          description: Internal company job identifier
          example: "2024-01-15-stripe-software-engineer-abc123"
        title:
          type: string
          description: Job title
          example: "Software Engineer"
        company:
          type: string
          description: Company name
          example: "Stripe"
        location:
          type: string
          nullable: true
          description: Job location
          example: "San Francisco, CA"
        description:
          type: string
          nullable: true
          description: Job description
        external_link:
          type: string
          nullable: true
          description: External job posting URL
          example: "https://jobs.stripe.com/abc123"
        source:
          type: string
          nullable: true
          description: Source of the job posting
          example: "Company Site"
        tags:
          type: array
          items:
            type: string
          description: Job tags
          example: ["engineering", "backend", "full-time"]
        scraped_at:
          type: string
          format: date-time
          description: When the job was scraped
          example: "2024-01-15T10:30:00.000Z"
        created_at:
          type: string
          format: date-time
          description: When the job was created in the system
          example: "2024-01-15T10:30:00.000Z"
        updated_at:
          type: string
          format: date-time
          description: When the job was last updated
          example: "2024-01-15T10:30:00.000Z"
      required:
        - id
        - title
        - company
        - scraped_at
        - created_at
        - updated_at
    JobCreate:
      type: object
      required:
        - title
        - company
      properties:
        title:
          type: string
          description: Job title
          minLength: 1
          example: "Software Engineer"
        company:
          type: string
          description: Company name
          minLength: 1
          example: "Tech Corp"
        location:
          type: string
          nullable: true
          description: Job location
          example: "San Francisco, CA"
        description:
          type: string
          nullable: true
          description: Job description
          example: "We are looking for a talented software engineer..."
        external_link:
          type: string
          nullable: true
          description: External job posting URL
          format: uri
          example: "https://jobs.techcorp.com/software-engineer"
        internal_id:
          type: string
          nullable: true
          description: Internal company job identifier
          example: "ENG-2024-001"
        source:
          type: string
          nullable: true
          description: Source of the job posting
          example: "Company Website"
        tags:
          type: array
          items:
            type: string
          description: Job tags
          example: ["engineering", "full-time", "remote"]
    JobImport:
      type: object
      properties:
        internal_id:
          type: string
          nullable: true
          description: Internal company job identifier
        title:
          type: string
          description: Job title
        company:
          type: string
          description: Company name
        location:
          type: string
          nullable: true
          description: Job location
        description:
          type: string
          nullable: true
          description: Job description
        external_link:
          type: string
          nullable: true
          description: External job posting URL
        source:
          type: string
          nullable: true
          description: Source of the job posting
        scraped_at:
          type: string
          format: date-time
          description: When the job was scraped
      required:
        - title
        - company
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: "Invalid request parameters"
        stack:
          type: string
          nullable: true
          description: Stack trace (only in development)
      required:
        - ok
        - error
...